// src/generated/model/academicYearRequest.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { AcademicYearRequestStatus } from './academicYearRequestStatus.ts';

export interface AcademicYearRequest {
  /**
     * @minLength 0
     * @maxLength 50
     */
  name: string;
  description?: string;
  startDate: string;
  endDate: string;
  status: AcademicYearRequestStatus;
  isCurrentYear?: boolean;
}