// src/generated/model/updateAcademicYearRequest.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { UpdateAcademicYearRequestStatus } from './updateAcademicYearRequestStatus.ts';

export interface UpdateAcademicYearRequest {
  /**
     * @minLength 0
     * @maxLength 50
     */
  name: string;
  description?: string;
  startDate: string;
  endDate: string;
  status: UpdateAcademicYearRequestStatus;
  isCurrentYear?: boolean;
}